main.py
import sys
import os
from google import genai
from google.genai import types
from dotenv import load_dotenv

from prompts import system_prompt

from functions.get_files_info import schema_get_files_info
from functions.get_files_content import schema_get_file_content
from functions.run_python import schema_run_python_file
from functions.write_file import schema_write_file


available_functions = types.Tool(
    function_declarations=[
        schema_get_files_info, schema_write_file, schema_get_file_content, schema_run_python_file
    ]
)

def main():
    load_dotenv()
    
    verbose = "--verbose" in sys.argv
    args = []
    for arg in sys.argv[1:]:
        if not arg.startswith("--"):
            args.append(arg)

    if not args:
        print("AI Code Assistant")
        print('\nUsage: python main.py "your prompt here" [--verbose]')
        print('Example: python main.py "How do I fix the calculator?"')
        sys.exit(1)

    api_key = os.environ.get("GEMINI_API_KEY")
    client = genai.Client(api_key=api_key)

    user_prompt = " ".join(args)

    if verbose:
        print(f"User prompt: {user_prompt}\n")

    messages = [
        types.Content(role="user", parts=[types.Part(text=user_prompt)]),
    ]
    for i in range(20):
        try:
            response = generate_content(client, messages, verbose)
            if response.text and not response.function_calls:
                print("Final response:")
                print(response.text)
                break
        except Exception as e:
            print(f"ERROR: {e}")
            break


def generate_content(client, messages, verbose):
    response = client.models.generate_content(
        model="gemini-2.0-flash-001",
        contents=messages,
        config=types.GenerateContentConfig(
            tools=[available_functions], system_instruction=system_prompt
        )
    )

    for candidate in response.candidates:
        messages.append(candidate.content)

    if verbose:
        print("Prompt tokens:", response.usage_metadata.prompt_token_count)
        print("Response tokens:", response.usage_metadata.candidates_token_count)
    print("Response:")
    if response.function_calls:
        for function_call_part in response.function_calls:
            function_call_result = call_function(function_call_part, verbose=verbose)
            
            try:
                response_data = function_call_result.parts[0].function_response.response
                    # If we get here, it exists

                if verbose:  
                    print(f"-> {response_data}")

                messages.append(function_call_result)

            except (IndexError, AttributeError):
                raise Exception("Invalid function call result structure")
    else:
        # Only print text if no function calls
        print(response.text)
    return response


from functions.write_file import write_file
from functions.get_files_content import get_file_content
from functions.get_files_info import get_files_info
from functions.run_python import run_python_file


def call_function(function_call_part, verbose=False):
    
    if verbose:
        print(f"Calling function: {function_call_part.name}({function_call_part.args})")
    else:
        print(f" - Calling function: {function_call_part.name}")

    function_map = {
        "run_python_file": run_python_file,
        "get_files_info": get_files_info,
        "write_file": write_file,
        "get_file_content": get_file_content
    }

    function_call_part.args["working_directory"] = "./calculator"

    if function_call_part.name in function_map:
        function_result = function_map[function_call_part.name](**function_call_part.args)
        return types.Content(
            role="tool",
            parts=[
                types.Part.from_function_response(
                    name=function_call_part.name,
                    response={"result": function_result},
                )
            ],
        )
    else:
        return types.Content(
            role="tool",
            parts=[
                types.Part.from_function_response(
                    name=function_call_part.name,
                    response={"error": f"Unknown function: {function_call_part.name}"},
                )
            ],
        )   

if __name__ == "__main__":
    main()
